[Andry] - fix button visibility

+++ resources/views/livewire/trd-tire1/transaction/sales-order/detail.blade.php
@@ -193,8 +193,10 @@
                                 loading="true" button-name="Cetak Surat Jalan" iconPath="print.svg"
                                 enabled="true" />
                         @endif
-                        <x-ui-button clickEvent="delete" :action="$actionValue" :enabled="$isDeliv ? 'false' : ($canUpdateAfterPrint ? 'true' : 'false')"
-                            type="delete" enableConfirmationDialog="true" :permissions="$permissions" />
+                        @if ($this->canShowDeleteButton())
+                            <x-ui-button clickEvent="delete" :action="$actionValue" :enabled="$isDeliv ? 'false' : ($canUpdateAfterPrint ? 'true' : 'false')"
+                                type="delete" enableConfirmationDialog="true" :permissions="$permissions" />
+                        @endif
 
                         <x-ui-button clickEvent="Save" :action="$actionValue" type="save" :enabled="$canUpdateAfterPrint || $canSaveButtonEnabled ? 'true' : 'false'" />
                     </x-ui-footer>
